requirements.txt
fastapi
uvicorn
pytest
appium-python-client
redis